backend/django/requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
astroid==2.12.10
asttokens==2.0.5
attrs==21.4.0
autopep8==1.7.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.10.0
bleach==4.1.0
certifi==2022.9.24
cffi==1.15.0
chardet==3.0.4
charset-normalizer==2.1.1
click==6.7
colorama==0.4.5
cryptography==38.0.1
DateTime==4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
distlib==0.3.6
distro==1.8.0
Django==3.2.13
django-background-tasks==1.2.5
django-binary-database-files==1.0.17
django-compat==1.0.15
django-cors-headers==3.13.0
django-environ==0.8.1
django-filter==22.1
djangorestframework==3.13.1
djongo==1.3.6
dnspython==2.2.1
docker==6.0.0
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
entrypoints==0.4
et-xmlfile==1.1.0
executing==0.8.3
filelock==3.8.0
gevent==21.12.0
gfm==0.0.3
greenlet==1.1.3
grequests==0.3.0
gunicorn==20.1.0
html2text==2015.11.4
idna==2.10
ipykernel==6.9.2
ipython==8.1.1
ipython-genutils==0.2.0
ipywidgets==7.7.0
isort==5.10.1
jedi==0.18.1
Jinja2==3.1.2
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.1.0
lazy-object-proxy==1.7.1
linecache2==1.0.0
lxml==4.9.0
Markdown==2.6.11
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mccabe==0.7.0
mistune==0.8.4
mongoengine==0.24.2
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
openpyxl==3.0.9
optional-django==0.1.0
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
paramiko==2.11.0
parso==0.8.3
pep8==1.7.1
pickleshare==0.7.5
pipenv==2022.10.9
platformdirs==2.5.2
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
py-gfm==0.1.1
pycodestyle==2.9.1
pycparser==2.21
Pygments==2.5.1
pyint==1.1.0
PyJWT==2.4.0
pylint==2.15.3
pymongo==3.12.3
PyNaCl==1.5.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.1
pywin32==304
pywinpty==2.0.5
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.2.2
QtPy==2.0.1
requests==2.28.1
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.2.4
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
texttable==1.6.4
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.5
tornado==6.1
traceback2==1.4.0
traitlets==5.1.1
typing_extensions==4.4.0
tzdata==2022.4
unittest2==1.1.0
urllib3==1.26.12
virtualenv==20.16.5
virtualenv-clone==0.5.7
watchdog==2.1.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.59.0
websocket-server==0.4
whitenoise==3.3.1
widgetsnbextension==3.6.0
wrapt==1.11.0
zope.event==4.5.0
zope.interface==5.4.0
